+++ test/index.ts
@@ -1,4 +1,4 @@
-import { describe, before } from 'mocha';
+import { before } from 'mocha';
 import { use as chaiUse } from 'chai';
 import chaiExclude from 'chai-exclude';
 import { setupEnv } from 'src/environment';
@@ -10,7 +10,5 @@ before(async () => {
   await run();
 });
 
-describe('Mutation', () => {
-  require('./create-user-mutation');
-  require('./login-mutation');
-});
+require('./mutation');
+require('./query');
