+++ test/mutation/index.ts
@@ -1,0 +1,6 @@
+import { describe } from 'mocha';
+
+describe('mutation', () => {
+  require('./create-user');
+  require('./login');
+});
