+++ test/query/index.ts
@@ -1,0 +1,5 @@
+import { describe } from 'mocha';
+
+describe('query', () => {
+  require('./user');
+});
