+++ tsconfig.json
@@ -13,5 +13,6 @@
     "forceConsistentCasingInFileNames": true,    
     "strict": true,                              
     "skipLibCheck": true                         
-  }
+  },
+  "include": ["./src/**/*.ts", "./test/**/*.ts"]
 }
